fix: restructure server actions folder structure
