+++ components/FileUpload/FileUpload.tsx
@@ -7,14 +7,13 @@ import Image from 'next/image';
 import { toast } from '../ui/use-toast';
 import { OurFileRouter } from '@/app/api/uploadthing/core';
 import { DeleteImageDialog } from '../index';
-import useBeforeUnload from '@/hooks/useBeforeUnLoad';
 
 type FileUploadProps = {
+	imagePreview:ImageResult|null
 	setImage: Dispatch<SetStateAction<ImageResult | null>>;
 };
 
-export default function FileUpload({ setImage }: FileUploadProps) {
-	const { imagePreview, setImagePreview } = useBeforeUnload();
+export default function FileUpload({ setImage, imagePreview }: FileUploadProps) {
 
 	return (
 		<div className='h-full w-full max-w-[400px]'>
@@ -29,7 +28,7 @@ export default function FileUpload({ setImage }: FileUploadProps) {
 					/>
 					<DeleteImageDialog
 						key={imagePreview.key}
-						setImagePreview={setImagePreview}
+						setImagePreview={setImage}
 					/>
 				</>
 			) : (
@@ -37,7 +36,6 @@ export default function FileUpload({ setImage }: FileUploadProps) {
 					className='min-h-[400px] w-full rounded-2xl border-0 text-black shadow-xl shadow-white-100 ut-button:mt-5 ut-button:rounded-md ut-button:bg-primary-light ut-button:px-5 ut-button:text-xs ut-button:text-white dark:shadow-black-100 ut-button:dark:bg-primary-dark ut-allowed-content:dark:text-white ut-label:dark:text-white ut-upload-icon:dark:text-white'
 					endpoint='imageUploader'
 					onClientUploadComplete={(res) => {
-						setImagePreview(res ? res[0] : null);
 						setImage(res ? res[0] : null);
 						toast({
 							className: 'text-white font-normal bg-green-600',
