+++ components/ReviewForm/ReviewForm.tsx
@@ -1,19 +1,17 @@
 'use client';
 
 import { useState } from 'react';
-import { useRouter } from 'next/navigation';
 
 import { toast } from '../ui/use-toast';
 import { Input } from '../ui/input';
 import { Button } from '@/components/ui/button';
 import { Textarea } from '../ui/textarea';
 import { FileUpload } from '../index';
-import { postReview } from '@/sanity/actions/review';
 import { formReviewFields } from '@/constant';
+import { postReview } from '@/serveractions';
 
 export default function ReviewForm() {
 	const [image, setImage] = useState<ImageResult | null>(null);
-	const router = useRouter();
 
 	const handleSubmit = async (data: FormData) => {
 		try {
@@ -27,7 +25,7 @@ export default function ReviewForm() {
 			await postReview(data, image);
 
 			toast({ title: 'Success' });
-			router.push('/', { scroll: true });
+			window.location.href = '/';
 		} catch (error) {
 			toast({
 				className: 'text-white',
@@ -43,7 +41,7 @@ export default function ReviewForm() {
 				action={handleSubmit}
 				className='flex h-full w-full flex-wrap items-center justify-center gap-20 '
 			>
-				<FileUpload setImage={setImage} />
+				<FileUpload setImage={setImage} imagePreview={image} />
 				<div className='h-full w-full  max-w-sm'>
 					{formReviewFields.map((formField) => (
 						<div key={formField.label} className='space-y-5'>
