+++ serveractions/index.ts
@@ -1,0 +1,6 @@
+'use server';
+
+import { deleteImage } from './uploadthing.actions';
+import { postReview } from './sanity.actions';
+
+export { deleteImage, postReview };
