+++ serveractions/uploadthing.actions.ts
@@ -1,5 +1,3 @@
-'use server';
-
 import { utapi } from 'uploadthing/server';
 
 export const deleteImage = async (key: string) => {
